% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_flat_cosine}
\alias{create_flat_cosine}
\title{Create an IndexFlatIP (inner-product) index for cosine similarity (use with normalized vectors)}
\usage{
create_flat_cosine(d)
}
\arguments{
\item{d}{Integer, vector dimension.}
}
\value{
An rfaiss index handle (python-backed).
}
\description{
Create an IndexFlatIP (inner-product) index for cosine similarity (use with normalized vectors)
}
