% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_flat_l2}
\alias{create_flat_l2}
\title{Create a FlatL2 index}
\usage{
create_flat_l2(d)
}
\arguments{
\item{d}{Integer, vector dimension.}
}
\value{
An rfaiss index handle (python-backed).
}
\description{
Create a FlatL2 index
}
