% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{index_add}
\alias{index_add}
\title{Add vectors to an index (L2)}
\usage{
index_add(handle, x)
}
\arguments{
\item{handle}{Object returned by \code{\link[=create_flat_l2]{create_flat_l2()}}.}

\item{x}{Numeric matrix (rows = vectors, ncol must equal index dimension).}
}
\description{
Add vectors to an index (L2)
}
