% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{index_search}
\alias{index_search}
\title{Search k nearest neighbors (L2)}
\usage{
index_search(handle, q, k = 5L)
}
\arguments{
\item{handle}{rfaiss index handle.}

\item{q}{Query matrix (rows = queries, ncol must match index dimension).}

\item{k}{Number of neighbors to return.}
}
\value{
A list with \code{distances} and \code{indices} matrices.
}
\description{
Search k nearest neighbors (L2)
}
