% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{load_index}
\alias{load_index}
\title{Load a FAISS index from disk}
\usage{
load_index(path)
}
\arguments{
\item{path}{Path created by \code{\link[=save_index]{save_index()}}.}
}
\value{
rfaiss index handle.
}
\description{
Load a FAISS index from disk
}
