% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerank_helpers.R
\name{mmr_select}
\alias{mmr_select}
\title{MMR reranking}
\usage{
mmr_select(query_emb, candidate_embs, lambda = 0.5, top_n = 5)
}
\arguments{
\item{query_emb}{numeric vector (d)}

\item{candidate_embs}{matrix (n x d)}

\item{lambda}{tradeoff 0..1}

\item{top_n}{number of output indices to pick (<= n)}
}
\value{
integer vector of selected indices (1-based, into candidate_embs)
}
\description{
MMR reranking
}
