% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfaiss-package.R
\docType{package}
\name{rfaiss-package}
\alias{rfaiss}
\alias{rfaiss-package}
\title{rfaiss: FAISS from R}
\description{
Lightweight R API that talks to FAISS (Python backend for now).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
