% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{save_index}
\alias{save_index}
\title{Save a FAISS index to disk}
\usage{
save_index(handle, path)
}
\arguments{
\item{handle}{rfaiss index handle.}

\item{path}{File path (e.g., "myindex.faiss").}
}
\description{
Save a FAISS index to disk
}
