% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faiss_py.R
\name{use_rfaiss_python}
\alias{use_rfaiss_python}
\title{Point rfaiss to a specific Python environment (optional)}
\usage{
use_rfaiss_python(env)
}
\description{
Point rfaiss to a specific Python environment (optional)
}
